requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
lxml==4.9.2
outcome==1.2.0
packaging==23.1
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
python-telegram-bot==20.3
requests==2.29.0
schedule==1.2.0
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
telegram==0.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.16
webdriver-manager==3.8.6
wsproto==1.2.0
youtube-transcript-api==0.6.0
